Prevent deleting UserSongVote record
Remove post_delete signal in favor of overring model method
